requirement.txt
Flask
transformers
torch
Pillow
